move videoPlayerView to study/relay
